refactor: rewire analysis imports and tests to use analysis.qc subpackage
